requirements.txt
Flask
cryptography
gunicorn
ruamel.yaml